decomp progress: WiiProfileMgr::AddIdToDeleteQueue 0.9%->100%
- WiiProfileMgr::AddIdToDeleteQueue 0.9%->100% (queue scan + Handle msg)
- Multiple incremental SDK/network edits

+++ src/sdk/RVL_SDK/revolution/sc/scsystem.h
@@ -111,7 +111,7 @@ typedef struct SCControl {
 } SCControl;
 
 void SCInit(void);
-u32 SCCheckStatus(void);
+u8 SCCheckStatus(void);
 
 BOOL SCFindByteArrayItem(void* dst, u32 len, SCItemID id);
 BOOL SCReplaceByteArrayItem(const void* src, u32 len, SCItemID id);
